upd(brave-{browser,keyring}-deb): `1.66.118` -> `1.70.123`, `1.13` -> `1.17` (#6498)
- `brave-browser-deb` wasn't updated in 4 months. I constantly use Brave
browser, so I'm interested in keeping this package up to date and I
added myself as maintainer, I hope it's okay.
- Added `arm64` support
- I don't know if there's a reason to use github links for deb sources
over everything else, but I can only speak for the opposite: for some
people (myself included, unfortunately) downloading from github is
usually painfully slow, even tho generally network speed is okay. For
example
[here](https://reddit.com/r/github/comments/ekvvff/extremely_slow_downloads_from_github/)
are complains about it from 5 to 2 years ago, and I know that these are
from a long time ago and that's just a bunch of people, but the fact
that I still experience the same problem makes me think that there still
may be other pacstall users wanting to download Brave but having the
same issue. And changing the source fixes it (in my case, ~30mins from
github vs <1min from brave deb repo). So I hope you'll find this change
justified. After all, that's also official Brave source and it was
already used in `brave-keyring-deb`.