add: `qt6-style-kvantum` (#6536)
## Progress 

- [x] Edit packagelist
- [x] Add initial pacscript
- [ ] Contact devs
- [x] Add maintainer to pacscript

This package
[exists](https://packages.ubuntu.com/oracular/qt6-style-kvantum) in
Ubuntu oracular (thus I used this name instead of updating unmaintained
`kvantum` package, which I suppose could be removed if such operation
could be done in this repo: if there still are packages that depend on
it, `qt5-style-kvantum | qt6-style-kvantum` could be used instead). But
required dependencies exist in Debian testing and sid, and in KDE Neon.
In Ubuntu repo the package is split in `${pkgname}`, `${pkgname}-themes`
and `${pkgname}-l10n`, but I kept everything in one package for user
convenience (users probably want l10n anyway, it's just a few Kb, and
themes take a few Mb, so I doubt it can be considered a bloat and so I
avoided extra work, but if you insist I will make a split package).