upd(telegram-bin): `4.14.9` -> `5.9.0` (#6816)
Many improvements are made based on the `telegram-desktop-bin` package
in the AUR. Most notably;
1. A new version, obviously.
2. Changed the URL to fetch the source AppImage from since the previous
one was taken down long ago.
3. Changes introduced by syncing the PKGBUILD from the AUR, most
prominently;
3.1. `.desktop` file to make the program appear in the apps
menu/launcher.
3.2. `.protocol` file for KDE.
3.3. Disablement of self-update mechanism since it won't work without
running Telegram as root or making the binary writable by anyone, both
of which is more dangerous than one another, due to the fact that it's
stored in `/usr/bin` anyway.

I tested the `pacscript` myself and confirmed it works, but you may
still want to check it anyway for great measure.

Co-authored-by: Bahar KURT <kurtbahartr@users.noreply.github.com>